# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SecuritySummary {
  pending: number;
  criticalAlerts: number;
  monitoredFiles: number;
  chainIntact: boolean;
  lastScan: string;
  scanInterval?: string;
}

export type SecurityAlertSeverity = typeof SecurityAlertSeverity[keyof typeof SecurityAlertSeverity];


export const SecurityAlertSeverity = {
  critical: 'critical',
  warning: 'warning',
  info: 'info',
} as const;

export interface SecurityAlert {
  id: string;
  kind: string;
  path: string;
  detail: string;
  severity: SecurityAlertSeverity;
  timestamp: string;
}

export type ProtectionRequestStatus = typeof ProtectionRequestStatus[keyof typeof ProtectionRequestStatus];


export const ProtectionRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  denied: 'denied',
} as const;

export interface ProtectionRequest {
  id: string;
  path: string;
  reason: string;
  status: ProtectionRequestStatus;
  createdAt: string;
  /** @nullable */
  decidedAt?: string | null;
  /** @nullable */
  decidedBy?: string | null;
  /** @nullable */
  encryptedTo?: string | null;
}

export interface DecisionInput {
  /** @minLength 2 */
  approver: string;
}

export type DecisionResultStatus = typeof DecisionResultStatus[keyof typeof DecisionResultStatus];


export const DecisionResultStatus = {
  approved: 'approved',
  denied: 'denied',
} as const;

export interface DecisionResult {
  status: DecisionResultStatus;
  message: string;
  /** @nullable */
  encryptedTo?: string | null;
}

export interface AuditEvent {
  id: string;
  timestamp: string;
  action: string;
  target: string;
  actor: string;
  detail: string;
}

export interface AuditTrail {
  chainIntact: boolean;
  events: AuditEvent[];
}