# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /security/requests/{id}/approve:
    post:
      operationId: approveProtectionRequest
      tags: [security]
      summary: Approve and protect a flagged file
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "200":
          description: Approved request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResult"
        "400":
          description: Missing approver
        "404":
          description: Request not found
  /security/requests/{id}/deny:
    post:
      operationId: denyProtectionRequest
      tags: [security]
      summary: Deny a flagged file
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "200":
          description: Denied request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResult"
        "400":
          description: Missing approver
        "404":
          description: Request not found
...
components:
  parameters:
    RequestId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    DecisionInput:
      type: object
      required: [approver]
      properties:
        approver: { type: string, minLength: 2 }
    DecisionResult:
      type: object
      required: [status, message]
      properties:
        status: { type: string, enum: [approved, denied] }
        message: { type: string }
        encryptedTo: { type: ["string", "null"] }
...