# lib/api-zod/src/generated/types/decisionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionResultStatus } from './decisionResultStatus';

export interface DecisionResult {
  status: DecisionResultStatus;
  message: string;
  /** @nullable */
  encryptedTo?: string | null;
}